Represent lines as u32 and columns as u16
With this change we represent source location lines as u32 and columns
as u16. Doing so is less wasteful than using machine words for both and
it is what both our GSYM and DWARF resolvers use internally. On top of
that doing so provides us with a few padding bytes inside the
symbolize::Sym type that we can use for other purposes, should the need
arise.

+++ src/gsym/resolver.rs
@@ -178,7 +178,7 @@ impl SymResolver for GsymResolver<'_> {
                 let path = Path::new(dirname).join(filename);
                 return Some(AddrLineInfo {
                     path,
-                    line: lntab_row.file_line as usize,
+                    line: lntab_row.file_line,
                     column: 0,
                 })
             }
